+++ shuup/admin/modules/categories/views/list.py
@@ -17,6 +17,7 @@ from shuup.core.models import Category, CategoryStatus, CategoryVisibility
 class CategoryListView(PicotableListView):
     model = Category
     default_columns = [
+        Column("image", _("Image"), sortable=False, linked=True, raw=True),
         Column(
             "name", _(u"Name"), sortable=False, display="format_name", linked=True,
             filter_config=MPTTFilter(