+++ shuup/admin/modules/services/__init__.py
@@ -52,6 +52,7 @@ class ServiceModule(AdminModule):
                 text=self.name,
                 url=self.menu_entry_url,
                 category=SETTINGS_MENU_CATEGORY,
+                subcategory="payment_shipping",
                 ordering=self.menu_ordering
             )
         ]