+++ shuup/notify/admin_module/__init__.py
@@ -61,7 +61,10 @@ class NotifyAdminModule(AdminModule):
             MenuEntry(
                 text=_("Notifications"), icon="fa fa-code",
                 url="shuup_admin:notify.script.list",
-                category=SETTINGS_MENU_CATEGORY, ordering=9, aliases=[_("Show notification scripts")]
+                category=SETTINGS_MENU_CATEGORY,
+                subcategory="store",
+                ordering=9,
+                aliases=[_("Show notification scripts")]
             )
         ]